Move superblock features from Tally to Quorum classes
The superblock index doesn't really belong in the tally area. This moves
it to the quorum code where it fits in more naturally.
